+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- `cookie-banner` component
 - Sharing popover button to the discover search page
 - Whitelist functionality for preprint discover page